noisy_mnist/requirements.txt
absl-py==0.11.0
aiohttp==3.6.2
appdirs==1.4.3
array2gif==1.0.4
asciimatics==1.11.0
astroid==2.3.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
black==19.3b0
bleach==3.2.1
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.12.5
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.4.1
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
contextlib2==0.5.5
coverage==5.1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.23
cytoolz==0.10.1
dask==2.16.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distro==1.5.0
docker-pycreds==0.4.0
dvc==0.54.1
entrypoints==0.3
faiss==1.6.3
filelock==3.0.12
flatbuffers==1.12
funcy==1.15
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.9
google==2.0.3
google-api-core==1.22.4
google-auth==1.22.1
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grandalf==0.6
graphviz==0.16
grpcio==1.32.0
gym==0.18.0
gym-retro==0.8.0
gym-super-mario-bros==7.3.2
h5py==2.10.0
humanize==3.0.1
idna==2.9
imagecodecs-lite==2019.12.3
imageio==2.8.0
imageio-ffmpeg==0.4.3
importlib-metadata==2.0.0
inferno-pytorch==0.4.0
inflect==4.1.0
iniconfig==1.1.1
ipdb==0.13.7
ipytest==0.9.1
ipython==7.23.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lie-learn==0.0.1.post1
llvmlite==0.34.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.6.0
moviepy==1.0.3
multidict==4.7.5
nanotime==0.5.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nes-py==8.1.6
nest-asyncio==1.4.3
networkx==2.4
notedown==1.5.1
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.1
pandas==1.0.1
pandoc-attributes==0.1.7
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.8.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.1
ply==3.11
pooch==1.1.0
POT==0.7.0
proglog==0.1.9
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.13.0
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyfiglet==0.8.post1
pyglet==1.5.0
Pygments==2.6.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pyssht==1.3.4
pytest==6.2.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-mnist==0.7
-e git+https://github.com/KevinMusgrave/pytorch-metric-learning.git@bbb33f99c003f7a306d94e7763074387b8fe4a51#egg=pytorch_metric_learning
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
ray==0.8.4
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.15.100
s2cnn==1.0.0
schema==0.7.3
scikit-image==0.17.1
scikit-learn==0.22.2.post1
scipy==1.4.1
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
soupsieve==2.0
subprocess32==3.5.4
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
testpath==0.4.4
tifffile==2020.5.7
toml==0.10.2
toolz==0.10.0
torch==1.4.0
torchvision==0.2.1
torchviz==0.0.2
tornado==6.0.4
tqdm==4.59.0
traitlets==4.3.3
treelib==1.6.1
trimesh==3.8.4
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.9
wandb==0.10.27
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zc.lockfile==2.0
zipp==3.3.0
